+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/ImpTest.hs
@@ -26,6 +26,7 @@ module Test.Cardano.Ledger.Conway.ImpTest (
   submitGovAction_,
   submitGovActions,
   submitProposal,
+  submitAndExpireProposalToMakeReward,
   submitProposal_,
   submitProposals,
   submitFailingProposal,
@@ -42,6 +43,7 @@ module Test.Cardano.Ledger.Conway.ImpTest (
   registerDRep,
   setupSingleDRep,
   setupPoolWithStake,
+  setupPoolWithoutStake,
   conwayModifyPParams,
   getProposals,
   getEnactState,
@@ -102,6 +104,7 @@ module Test.Cardano.Ledger.Conway.ImpTest (
   cantFollow,
   getsPParams,
   currentProposalsShouldContain,
+  setupDRepWithoutStake,
 ) where
 
 import Cardano.Crypto.DSIGN (DSIGNAlgorithm (..), Ed25519DSIGN, Signable)
@@ -335,6 +338,34 @@ registerDRep = do
   dreps `shouldSatisfy` Map.member (KeyHashObj khDRep)
   pure khDRep
 
+-- | In contrast to `setupSingleDRep`, this function does not make a UTxO entry
+-- that could count as delegated stake to the DRep, so that we can test that
+-- rewards are also calculated nonetheless.
+setupDRepWithoutStake ::
+  forall era.
+  ( ConwayEraTxCert era
+  , ShelleyEraImp era
+  ) =>
+  ImpTestM
+    era
+    ( KeyHash 'DRepRole (EraCrypto era)
+    , KeyHash 'Staking (EraCrypto era)
+    )
+setupDRepWithoutStake = do
+  drepKH <- registerDRep
+  delegatorKH <- freshKeyHash
+  deposit <- getsNES $ nesEsL . curPParamsEpochStateL . ppKeyDepositL
+  submitTxAnn_ "Delegate to DRep" $
+    mkBasicTx mkBasicTxBody
+      & bodyTxL . certsTxBodyL
+        .~ SSeq.fromList
+          [ mkRegDepositDelegTxCert @era
+              (KeyHashObj delegatorKH)
+              (DelegVote (DRepCredential $ KeyHashObj drepKH))
+              deposit
+          ]
+  pure (drepKH, delegatorKH)
+
 -- | Registers a new DRep and delegates the specified amount of ADA to it.
 setupSingleDRep ::
   forall era.
@@ -405,6 +436,28 @@ setupPoolWithStake delegCoin = do
               (pp ^. ppKeyDepositL)
           ]
   pure (khPool, credDelegatorPayment, credDelegatorStaking)
+
+setupPoolWithoutStake ::
+  (ShelleyEraImp era, ConwayEraTxCert era) =>
+  ImpTestM
+    era
+    ( KeyHash 'StakePool (EraCrypto era)
+    , Credential 'Staking (EraCrypto era)
+    )
+setupPoolWithoutStake = do
+  khPool <- registerPool
+  credDelegatorStaking <- KeyHashObj <$> freshKeyHash
+  deposit <- getsNES $ nesEsL . curPParamsEpochStateL . ppKeyDepositL
+  submitTxAnn_ "Delegate to stake pool" $
+    mkBasicTx mkBasicTxBody
+      & bodyTxL . certsTxBodyL
+        .~ SSeq.fromList
+          [ RegDepositDelegTxCert
+              credDelegatorStaking
+              (DelegStake khPool)
+              deposit
+          ]
+  pure (khPool, credDelegatorStaking)
 
 -- | Submits a transaction with a Vote for the given governance action as
 -- some voter
@@ -585,6 +638,25 @@ trySubmitGovAction ::
 trySubmitGovAction ga = do
   let mkGovActionId tx = GovActionId (txIdTx tx) (GovActionIx 0)
   fmap mkGovActionId <$> trySubmitGovActions (pure ga)
+
+submitAndExpireProposalToMakeReward ::
+  ConwayEraImp era =>
+  Int ->
+  Credential 'Staking (EraCrypto era) ->
+  ImpTestM era ()
+submitAndExpireProposalToMakeReward expectedReward stakingC = do
+  rewardAccount <- getRewardAccountFor stakingC
+  EpochInterval lifetime <- getsNES $ nesEsL . curPParamsEpochStateL . ppGovActionLifetimeL
+  gai <-
+    submitProposal $
+      ProposalProcedure
+        { pProcDeposit = Coin $ fromIntegral expectedReward
+        , pProcReturnAddr = rewardAccount
+        , pProcGovAction = TreasuryWithdrawals mempty def
+        , pProcAnchor = def
+        }
+  passNEpochs $ 2 + fromIntegral lifetime
+  expectMissingGovActionId gai
 
 -- | Submits a transaction that proposes the given governance action
 trySubmitGovActions ::